refactor!: rearrange functions to be declared in a similar order to the order they're run
Summary: I don't like how rn each function is written in nearly the opposite
order to the order we run them.  It is not very "clean code".  This is done
because in Lua, local functions are not hoisted so I cannot use a local
function before I declare it.  Additionally, I wanted to avoid making a bunch
of global functions (which are hoisted) to be polite to our users.

I worked around this limitation a bit by adding many nested functions, and only
having a few global functions.

I also deleted a test in order to move the function that test was testing into another function

Test plan:
```sh
$ plugin/open-remote-test #all passed
```